bug(dynamic navbar links): make links dynamic on logout
delete registerActions.js
move actions for registering to sessionActions.js
delete the registerApi file
move the api method for registering users to the sessionApi class
delete the registerReducer file
move all reducers for registering to the sessionReducer filer
delete the register object in initial state
modify tests for register actions
move tests for register reducers to tests for session reducers
remove the register key from the root reducer
modify the Header component to show the user's name when logged in
modify tests for the Signup component
modify tets for the Header component
modify the Signup component to use sessionActions

[finishes #158284357]
