+++ client/actions/sessionActions.js
@@ -52,3 +52,30 @@ export function logOutUser() {
     }
   };
 }
+
+export function registerSuccess(response) {
+  return { type: types.REGISTER_SUCCESS, response };
+}
+
+export function registerFailure(response) {
+  return { type: types.REGISTER_FAILURE, response };
+}
+
+export function registerUser(credentials) {
+  return (dispatch) => {
+    dispatch(loginRequest());
+    return sessionApi.register(credentials)
+      .then((response) => {
+        const decodedToken = decode(response.data.token);
+        sessionStorage.setItem('jwt', response.data.token);
+        sessionStorage.setItem('userId', decodedToken.id);
+        sessionStorage.setItem('name', decodedToken.name)
+        dispatch(registerSuccess(response.data));
+        return response.data.message;
+      })
+      .catch((error) => {
+        dispatch(registerFailure(error.data));
+        throw error.data.message;
+      });
+  };
+}
